refactor Server.ServeCodec to use ServeRequest
This refactor removes the ability to handle multiple requests at once,
but ensures that tests use ServeRequest.

Consul only uses ServeRequest, so while this change removes functionality, it makes the
tests more relevant for the one use case we care about.

This change is being made to reduce the exported interface of the package
and to remove the need to implement the interceptor in ServeCodec as well.