Move functions not used by Consul to be test helpers
These functions for running a net/rpc server from within an http.Server
are used to test net/rpc, but are never used by Consul.

By moving these to be test helpers we remove the need to instrument them
with the new interceptor. We also reduce the code that we need to test
while making changes to net/rpc.